telegram_payment_bot/requirements.txt
python-telegram-bot
SQLAlchemy
alembic
python-dotenv
apscheduler
